Allow for resinOS releases containing the '+' character in the release version name (e.g. 2.0.6+rev1.prod)
The first change is related to the fact that '+' is a special char
in a regex and we need to escape it to actually use it as the '+' char.

The second change addresses the fact the AWS S3 treats '+' in the
path section of an URL as a space character. In our case it is part
of a actual directory name in the URL's path part so we need S3 to
interpret the '+' char as the actual '+' char and not the space char,
therefore we encode it here using the equivalent hex code '%2B' used
for the '+' character.

Signed-off-by: Florin Sarbu <florin@resin.io>