run.sh: Fix module directory path
The module folder path was wrong. The module is built in a folder of
the format example_module_device-type_os-version
"example_module_intel-nuc_2.29.0+rev1.prod"

RESIN_HOST_OS_VERSION is "balenaOS 2.26.0+rev3"

Fix parsing to remove balenaOS and the space.

The modules are built with .dev/.prod folder suffix. This information
is not available inside the container. Use a delimiter * to make the
module load. This loads dev/prod both if both are built and in the
folder which is not ideal but better than the broken path we have
currently

Changelog-entry: Fix module directory path in run.sh
Signed-off-by: Zubair Lutfullah Kakakhel <zubair@balena.io>