workarounds.sh: Add workarounds for various device
We have been keeping the workarounds in separate branches but that is
harder for testing and tricky to manage. Fixes to master dont reach the
branches.

This commit adds a workaround.sh file which is called during the build.
Workarounds for various devices go here.

Changelog-entry: Add workarounds for various devices
Signed-off-by: Zubair Lutfullah Kakakhel <zubair@balena.io>