Workaround for up-board
This workaround re-compiles the module generating tools with flags
specific to the builder server CPU arch.

Otherwise, the tools can't be executed on the server because they are
compiled with flags specific to up-board's CPU version.

This error is avoided by applying the workaround:
" /bin/sh: 1: scripts/basic/fixdep: not found"

Signed-off-by: Sebastian Panceac <sebastian@balena.io>