run.sh: build modules at runtime
Move the module build steps to a runtime script.

Although this often takes much longer since it is running on the
target hardware, it can be worthwhile to have the device type
and os version pulled from the running image.

Change-type: major
Changelog-entry: move the module build steps to runtime
Signed-off-by: Kyle Harding <kyle@balena.io>