Add compatibility with UDEV=on
Some kernel modules may create devices under `/dev`, in which case the
recommended way to show those devices in the container is to set
`UDEV=on`. Since the dockerfile in this example repo replaces the
entrypoint, then that feature won't be available by default.

This updates the run.sh entrypoint to pass the CMD to `/usr/bin/entry.sh`
to make it compatible with that feature.

It also adds a disclaimer about modules that need to load in order.

Change-type: minor